pyproject.toml
[tool.poetry]
name = "trading-network"
version = "0.1.0"
description = ""
authors = ["ssher <ssher250110@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
django-filter = "^24.3"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
